requirements.txt
matplotlib.pyplot
pandas 
seaborn
streamlit
plotly.express